+++ packages/stream_video/lib/src/call/session/call_session.dart
@@ -624,7 +624,7 @@ class CallSession extends Disposable {
   void leave({String? reason}) {
     _logger.d(() => '[leave] reason: $reason');
     _isLeavingOrClosed = true;
-    _tracer.trace('call.leave', reason);
+    _tracer.trace('call.leaveReason', reason);
     sfuWS.leave(sessionId: sessionId, reason: reason);
   }